backend/requirements.txt
asyncpg~=0.30.0
annotated-types~=0.7.0
anyio~=4.7.0
python-decouple~=3.8
fastapi~=0.115.6
idna~=3.10
prometheus_client~=0.21.1
pydantic~=2.10.4
pydantic_core~=2.27.2
python-json-logger~=2.0.7
sniffio~=1.3.1
SQLAlchemy~=2.0.36
starlette~=0.41.3
typing_extensions~=4.12.2
uvicorn~=0.28.0
pandas~=2.2.1
greenlet~=3.1.1
qdrant-client==1.9.1